[FEATURE] user 메인페이지 포스터등록
팝업 번호와 포스터 이미지 번호를 매핑시키기

+++ backend/src/main/java/com/ohgiraffers/poppop/favorite/controller/FavoriteController.java
@@ -49,7 +49,14 @@ public class FavoriteController {
 
     // 찜목록 삭제
     @DeleteMapping("/favorite")
-    public ResponseEntity<?> deleteFavoritePopup(@RequestParam int popupNo,@RequestParam String id){
+    public ResponseEntity<?> deleteFavoritePopup(@RequestParam int popupNo, @AuthenticationPrincipal UserDetails userDetails){
+
+        if (userDetails == null) {
+            return ResponseEntity.status(401).body("Unauthorized");
+        }
+        String id = userDetails.getUsername();
+
+
 
         favoriteService.deleteFavorite(popupNo,id);
